Revert "abbreviated months"
This reverts commit f3c4ec34e148046a78255d67ef6f2116bdad4b2e.

+++ app/blog/utils.ts
@@ -76,9 +76,8 @@ export function formatDate(date: string, includeRelative = false) {
     formattedDate = 'Today'
   }
 
-  // Use 'short' for abbreviated month names
   let fullDate = targetDate.toLocaleString('en-us', {
-    month: 'short', // Abbreviated month names (e.g., Dec)
+    month: 'long',
     day: 'numeric',
     year: 'numeric',
   })
